Support Go modules
This adds support for Go modules.  Because this package is already
at version 2.4, we can't just directly add `go.mod` and `go.sum`;
instead, we need to bump the version to 3.0 and namespace the module
with `/v3`.  Remember, Go modules require a `/vN` namespace for every
version greater than 1.  Because we're already midway through version
2, I think that the best course of action would be to introduce Go
modules in a new major version (3.0).